Arduino Due HTTPS example
This example was done in Arduino IDE 1.18 and is used with Arduino Due and Ethernet Shield (W5100). To use simply copy the contents to Arduino folder in user profile Documents folder and launch the EthernetHTTPS.ino - don't forget do update the certificates in trust_anchors.h as said in the video.